Search by username on transaction view#4341
