+++ auth-web/src/models/searchfilter.ts
@@ -3,6 +3,6 @@ export interface SearchFilterParam {
   id: string
   placeholder: string
   labelKey: string
-  appliedFilterValue: string
-  filterInput: string
+  appliedFilterValue: any
+  filterInput: any
 }
